+++ pihello/__init__.py
@@ -1,3 +1,3 @@
 from .console import Console
 
-__version__ = "0.2.6"
+__version__ = "1.0.0"